[Decomp][Spell] Cohesion split: extract pre-cast validation to SpellChecks.cpp
Move the Spell:: pre-cast validation suite (CheckCast, CheckPetCast,
CheckCasterAuras, CanAutoCast, CheckRange, CalculatePowerCost, CheckPower,
IgnoreItemRequirements, CheckItems) out of Spell.cpp into a new SpellChecks.cpp
TU. Contiguous block, bodies only; declarations stay in Spell.h, so call sites
and linkage are unchanged. No file-statics or file-local helper classes in the
moved range. Spell.cpp 10806 -> 7842. Pure relocation, no behaviour change.
